# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GrowthAgent API",
    "version": "1.0.0",
    "description": "Autonomous AI growth system with ONE deterministic gatekeeper. Built for Razorpay AI Buildathon."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development & demo server"
    }
  ],
  "paths": {
    "/v1/carts/proposals": {
      "post": {
        "summary": "Create Cart Proposal",
        "description": "Submit shopping intent from buyer agent; starts asynchronous pipeline.",
        "security": [
          {
            "AgentKeyAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Proposal accepted; returns tx_id and stream/poll URLs."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthorized or revoked key."
          },
          "409": {
            "description": "Idempotency conflict."
          },
          "429": {
            "description": "Rate limited."
          }
        }
      }
    },
    "/v1/carts/proposals/{txId}": {
      "get": {
        "summary": "Poll Cart Proposal Status",
        "description": "Poll transaction state; returns terminal outcome when complete.",
        "security": [
          {
            "AgentKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction state (pending or terminal outcome union)."
          },
          "404": {
            "description": "Transaction not found or foreign agent."
          },
          "429": {
            "description": "Source exhausted its FAILED-authentication budget (RATE_LIMITED_HTTP). Successful polling is never throttled."
          }
        }
      }
    },
...
    "/v1/stream/{txId}": {
      "get": {
        "summary": "Server-Sent Events Transaction Audit Stream",
        "description": "SSE stream projecting the hash-chained audit log for one transaction. Admission is checked BEFORE the first byte, so a refusal is an ordinary JSON error rather than a half-open stream.",
        "responses": {
          "200": {
            "description": "SSE event stream."
          },
          "401": {
            "description": "Missing/invalid ticket or agent key."
          },
          "404": {
            "description": "Transaction not found or foreign agent."
          },
          "429": {
            "description": "Concurrent-stream ceiling reached (RATE_LIMITED_HTTP); details carry the global and per-agent limits."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      },
      "RazorpaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Razorpay-Signature"
      }
    }
  }
}